Review: MDEV-29639.
Implement an "unlikely" last_master_timestamp update block in the parallel
branch of the driver at its pre-scheduling time.
The lmt update block is entered under condition of
1. the very first event upon slave restart
2. a first event after wait for it in relay-log
3. delayed execution.

To the delayed mode, lmt needs an *estimation* at point to respect
the user expectation in the following.
At delaying the event`s execution SBM should not have (be displayed with) a value exceeding the CM`s delay parameter one.